refactor update score model into a component
